orders/src/events/listener/ticket-created-listener.ts
import { Listener, Subjects, TicketCreatedEvent } from "@tpthinh/common";
import { Message } from "node-nats-streaming";
import { queueGroupName } from "./queue-group-name";
import { Ticket } from "../../models/ticket";

export class TicketCreatedListener extends Listener<TicketCreatedEvent> {
  subject: Subjects.TicketCreated = Subjects.TicketCreated;
  queueGroupName = queueGroupName;

  async onMessage(
    { title, price, id }: TicketCreatedEvent["data"],
    msg: Message
  ) {
    const ticket = Ticket.build({
      id,
      title,
      price,
    });
    await ticket.save();

    msg.ack();
  }
}
